requirements.txt
aiohttp==3.9.1
python-dotenv==1.0.0
vkbottle==3.0.2
requests==2.31.0
pytz==2023.3
schedule==1.2.0